palantir: minor polish

+++ support/cas-server-support-consent-core/src/main/java/org/apereo/cas/consent/BaseConsentRepository.java
@@ -66,7 +66,7 @@ public abstract class BaseConsentRepository implements ConsentRepository {
         if (consent) {
             getConsentDecisions().remove(decision);
         } else {
-            decision.setId(RandomUtils.nextLong());
+            decision.setId(RandomUtils.nextInt());
         }
         getConsentDecisions().add(decision);
         return decision;
