+++ support/cas-server-support-wsfederation/src/main/java/org/apereo/cas/config/WsFederationAuthenticationConfiguration.java
@@ -71,7 +71,7 @@ class WsFederationAuthenticationConfiguration {
             final ConfigurableApplicationContext applicationContext) {
             return plan -> {
                 val service = new CasRegisteredService();
-                service.setId(RandomUtils.nextLong());
+                service.setId(RandomUtils.nextInt());
                 service.setEvaluationOrder(Ordered.HIGHEST_PRECEDENCE);
                 service.setName(service.getClass().getSimpleName());
                 service.setDescription("WS-Federation Authentication Request");
